+++ test/faker/person_test.exs
@@ -1,0 +1,10 @@
+defmodule Faker.PersonTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.Person
+  doctest Faker.Person.En
+  doctest Faker.Person.Es
+  doctest Faker.Person.Hy
+  doctest Faker.Person.It
+  doctest Faker.Person.PtBr
+end